A... moderate amount of refactoring
- Moved code into a code/src directory, outside of the actions directory
