+++ .github/code/src/dev/index.ts
@@ -1,0 +1,7 @@
+
+async function run(): Promise<void> {
+ 
+}
+
+run();
+
